requirements.txt
pillow
streamlit
fastapi
uvicorn
opencv-python
numpy
insightface
onnxruntime
python-multipart
